bluetooth: services: alexa_gadget: Move service and profile files
Move Alexa Gadgets profile to bluetooth subsystem,
such that it can be easily used by applications or samples
without having to copy files from the alexa_gadgets sample.
