+++ include/bluetooth/services/gadgets_profile.h
