+++ samples/bluetooth/alexa_gadget/CMakeLists.txt
@@ -13,7 +13,7 @@ target_sources(app PRIVATE
   src/main.c
 )
 
-add_subdirectory(src/gadgets)
+# Explicit linking needed to ensure protobuffer files are compiled before sample
 target_link_libraries(app PRIVATE ALEXA_GADGETS_PROFILE)
 
 # NORDIC SDK APP END
