--- samples/bluetooth/alexa_gadget/Kconfig
@@ -1,13 +1,0 @@
-#
-# Copyright (c) 2020 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-menu "Zephyr Kernel"
-source "Kconfig.zephyr"
-endmenu
-
-menu "Alexa Gadget"
-rsource "src/gadgets/Kconfig"
-endmenu
