+++ subsys/bluetooth/services/CMakeLists.txt
@@ -16,4 +16,4 @@ zephyr_sources_ifdef(CONFIG_BT_NUS_CLIENT nus_client.c)
 zephyr_sources_ifdef(CONFIG_BT_LBS lbs.c)
 zephyr_sources_ifdef(CONFIG_BT_LATENCY latency.c)
 zephyr_sources_ifdef(CONFIG_BT_LATENCY_CLIENT latency_client.c)
-zephyr_sources_ifdef(CONFIG_BT_GADGETS gadgets.c)
+add_subdirectory_ifdef(CONFIG_BT_GADGETS alexa_gadgets)
