--- subsys/bluetooth/services/Kconfig.gadgets
@@ -1,19 +1,0 @@
-#
-# Copyright (c) 2019 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-config BT_GADGETS
-	bool "Gadgets service"
-	select BT_NRF_SERVICES
-	help
-	  Enable Gadgets service.
-
-if BT_GADGETS
-
-module = BT_GADGETS
-module-str = GADGETS
-source "${ZEPHYR_BASE}/subsys/logging/Kconfig.template.log_config"
-
-endif # BT_GADGETS
