+++ subsys/bluetooth/services/alexa_gadgets/gadgets.c
