+++ subsys/bluetooth/services/alexa_gadgets/proto/custom_event.proto
